PS-5008: Memory leak in sync_latch_meta_init() after mysqld shutdown detected by ASan
Free all the resources acquired by InnoDB (mutexes, events, memory) in
case of a failed start (errors in
storage/innobase/srv/srv0start.cc::innobase_start_or_create_for_mysql())

Removed silencing leaks in tests

Added missing `innodb_init_abort()` in `innodb_init()`. It caused leaks

Added missing `recv_recovery_from_checkpoint_finish()` in `srv_start()`. It caused leaks

+++ storage/innobase/dict/dict0dict.cc
@@ -3921,6 +3921,8 @@ void dict_persist_init(void) {
 
 /** Clear the structure */
 void dict_persist_close(void) {
+  if (!dict_persist) return;
+
   ut::delete_(dict_persist->persisters);
 
 #ifndef UNIV_HOTBACKUP
